+++ gdsfactory/components/bend_circular.py
@@ -44,6 +44,7 @@ def bend_circular(
     ref = c << path
     c.add_ports(ref.ports)
 
+    c.absorb(ref)
     c.info["length"] = snap_to_grid(p.length())
     c.info["dy"] = float(abs(p.points[0][0] - p.points[-1][0]))
     c.info["radius"] = float(radius)
@@ -62,7 +63,6 @@ def bend_circular(
         for layer in layers_cladding or []:
             c.add_polygon(points, layer=layer)
 
-    c.absorb(ref)
     return c