+++ gdsfactory/routing/add_fiber_array.py
@@ -199,7 +199,8 @@ if __name__ == "__main__":
     # c = gf.components.mmi2x2()
     # c = gf.components.ring_single()
     # c = gf.components.straight_heater_metal()
-    c = gf.components.spiral(direction="NORTH")
+    # c = gf.components.spiral(direction="NORTH")
+    c = gf.components.mzi()
     cc = add_fiber_array(
         component=c,
         # optical_routing_type=0,