+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - add_pin_path now works with siepic
 - add add_pins_siepic in gf.add_pins
 - gf.path.extrude can also add pins
+- unpin `requirements.txt` [issue](https://github.com/gdsfactory/gdsfactory/issues/221)
 
 ## [4.3.6](https://github.com/gdsfactory/gdsfactory/pull/217)